server/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
channels==4.3.1
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
openai==2.8.1
pillow==12.0.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
